move: BrowserUtils into internal
